added menu files
+added it to the make file
