+++ source/game.c
@@ -96,19 +96,9 @@ void worldUpdate(Game *pGame){ // shifting the map
 
 }
 
-SDL_Texture* makeStringInToSDL_Texture(char string[NAME], SDL_Rect Pos, TTF_Font *pFont,SDL_Renderer* pRendererer){
-    char stringBuff[128];
-    snprintf(stringBuff,sizeof(stringBuff),"%s",string);
-    SDL_Color color ={255,255,255,255};
-    SDL_Surface* pSolidString = TTF_RenderText_Solid(pFont,stringBuff,color);
-    SDL_Texture* pTextureString = SDL_CreateTextureFromSurface(pRendererer,pSolidString);
-    SDL_FreeSurface(pSolidString);
-    return pTextureString;
-}
 
-void renderMenu(Game *pGame){
-    //SDL_RenderCopy(pGame->pRenderer,pGame->pMenu.)
-}
+
+
 
 
 int initialize_window(Game *pGame){ // Initialiserar SDL och skapar fönster
@@ -171,26 +161,6 @@ Camera *initialize_camera(void){
     return pCamera;
 }
 
-Menu* initialize_Menu(void){
-    Menu* pMenu = malloc(sizeof(Menu));
-    if(!pMenu){
-        fprintf(stderr,"Memory allocation failed for Menu\n");
-        return NULL;
-    }
-    pMenu->leter = 0;
-    char empty = {0};
-    strcpy(pMenu->stringPlayerName,&empty);
-    for (int i = 0; i < NUMMBER_OF_MENU_OPTIONS; i++){
-        pMenu->rect[i] = (SDL_Rect){0,0,0,0};   //fixar sen 
-    }
-    pMenu->pFont = TTF_OpenFont("resourses/RubikMaps-Regular.ttf",24);
-        if(!pMenu->pFont){
-        fprintf(stderr,"Error: Loding font: %s\n", TTF_GetError());
-        return false;
-    }
-    pMenu->open = true;
-    return pMenu;
-}
 
 void updateTileSize(Game *pGame){ 
     int width, height;
